requirements.txt
streamlit
pandas
numpy
scikit-learn
plotly
python-docx

